added network policy test
