+++ benchmarks/kubectl-mtb/test/benchmarks/create_network_policies/README.md
@@ -1,0 +1,12 @@
+# Create Network Policies <small>[MTB-PL2-BC-OPS-4] </small>
+**Profile Applicability:** 
+2
+**Type:** 
+Behavioral
+**Category:** 
+Self-Service Operations 
+**Description:** 
+Tenants should be able to perform self-service operations by creating own network policies in their namespaces. 
+**Remediation:**
+
+
